+++ threedi_modelchecker/tests/test_checks_other.py
@@ -4,6 +4,7 @@ from sqlalchemy.orm import aliased, Query
 from threedi_schema import constants, models
 
 from threedi_modelchecker.checks.other import (
+    ChannelManholeLevelCheck,
     ConnectionNodesDistance,
     ConnectionNodesLength,
     CrossSectionLocationCheck,
@@ -132,6 +133,64 @@ def test_open_channels_with_nested_newton(session):
 
     errors = check.get_invalid(session)
     assert len(errors) == 2
+
+
+channel_manhole_level_testdata = [
+    ("start", -1, -3, -2, 0),
+    ("start", -3, -1, -2, 1),
+    ("end", -3, -1, -2, 0),
+    ("end", -1, -3, -2, 1),
+]
+
+
+@pytest.mark.parametrize(
+    "manhole_location,starting_reference_level,ending_reference_level,manhole_level,errors_number",
+    channel_manhole_level_testdata,
+)
+def test_channel_manhole_level_check(
+    session,
+    manhole_location,
+    starting_reference_level,
+    ending_reference_level,
+    manhole_level,
+    errors_number,
+):
+    # using factories, create one minimal test case which passes, and one which fails
+    # once that works, parametrise.
+    # use nested factories for channel and connectionNode
+    starting_coordinates = "4.718300 52.696686"
+    ending_coordinates = "4.718255 52.696709"
+    start_node = factories.ConnectionNodeFactory(
+        the_geom=f"SRID=4326;POINT({starting_coordinates})"
+    )
+    end_node = factories.ConnectionNodeFactory(
+        the_geom=f"SRID=4326;POINT({ending_coordinates})"
+    )
+    channel = factories.ChannelFactory(
+        the_geom=f"SRID=4326;LINESTRING({starting_coordinates}, {ending_coordinates})",
+        connection_node_start=start_node,
+        connection_node_end=end_node,
+    )
+    # starting cross-section location
+    factories.CrossSectionLocationFactory(
+        the_geom="SRID=4326;POINT(4.718278 52.696697)",
+        reference_level=starting_reference_level,
+        channel=channel,
+    )
+    # ending cross-section location
+    factories.CrossSectionLocationFactory(
+        the_geom="SRID=4326;POINT(4.718264 52.696704)",
+        reference_level=ending_reference_level,
+        channel=channel,
+    )
+    # manhole
+    factories.ManholeFactory(
+        connection_node=end_node if manhole_location == "end" else start_node,
+        bottom_level=manhole_level,
+    )
+    check = ChannelManholeLevelCheck(nodes_to_check=manhole_location)
+    errors = check.get_invalid(session)
+    assert len(errors) == errors_number
 
 
 def test_node_distance(session):